moving actions to pdf actions
